Handle `sql.ErrNoRows` for GetAuthorization correctly.
Prior to this commit if the `sa.GetAuthorization` found no pending authz
rows **and** no authz rows for a given authz ID then the `sql.ErrNoRows`
result was wrapped in a `Rollback` and given to the caller. Since
`Rollback` wraps the error in its own type callers are unable to treat
the `sql.ErrNoRows` error as a no rows error and instead handle it as an
unexpected database error.

This commit changes `GetAuthorization` to treat `sql.ErrNoRows`
specially, returning it as-is.

This commit also removes the named return values since this is counter
to our current best practices. This function is getting a little messy
overall and could probably use a rewrite as a follow-up issue.

A unit test is included that failed prior to making the SA changed and
passes subsequently.