Only wrap Rollback error when tx.Rollback fails.
Prior to this commit the `Rollback` function always wrapped the provided
error in a `sa.RollbackError`. This makes it difficult for callers to
test the type of the original error (e.g. in the case of
`sql.ErrNoRows`). This commit updates the `Rollback` function to only
return a `sa.RollbackError` when the `tx.Rollback()` produces an error.