Add WrongAuthorizationState error code for UpdateAuthorization.
This commit adds a new boulder error type `WrongAuthorizationState`.
This error type is returned by the SA when `UpdateAuthorization` is
provided an authz that isn't pending. The RA and WFE are updated
accordingly such that this error percolates back through the API to the
user as a Malformed problem with a sensible description. Previously this
behaviour resulted in a ServerInternal error.