Treat DNAMEs as CNAMEs.
Since the legacy CAA spec does the wrong thing with CNAMEs (treating them as
CNAMEs), we do the same thing, essentially turning DNAMEs into CNAMEs. Note that
even though this change is made to the bdns package, it only takes effect when
the LegacyCAA flag is enabled, because that's the only time that VA looks at the
cname return value from LookupCAA.

Also, in the process, fix a bug in the previous LegacyCAA implementation.
Because the processing of records in LookupCAA was gated by `if
answer.Header().RRType == dnsType`, non-CAA responses were filtered out. This
wasn't caught by previous testing, because it was unittesting that mocked out
bdns.