Update `orderValid` to allow nil `CertificateSerial`
The NewOrder usage of `orderValid` means that `CertificateSerial` will
be nil but the order _is_ valid, just not finalized. This commit updates
the `grpc.orderValid` function to allow the nil `CertificateSerial`
field. The wrappers for the SA's `finalizeOrder` RPCs are updated to
additionally check that `orderValid` is true *and* `CertificateSerial`
is not nil.