using docker network with docker-compose (#3203)
Change docker-compose.yml to use a new docker network, not a legacy
bridge for networking.

For compatibility with current test setup entrypoint.sh adds
boulder-mysql to /etc/hosts in the boulder container. This avoids
dependency on Docker DNS proxy when running the tests.

The patch also fixes the volume declaration for the prom service to use
".", not "$PWD" to refer to the current directory as $PWD is not
defined by docker-compose at least in the version 1.16.1.

Document how to change test/config to use Boulder for clients running
in other docker containers.