Policy based issuance for wildcard identifiers.
This commit implements a first go at policy based issuance for wildcard
identifiers for the V2 API.

The policy component is updated to allow wildcard domains when
appropriate (e.g. in an order names field or in a CSR). The policy
component is also updated to know to only give a DNS-01 challenge for
an authorization corresponding to a wildcard identifier.

To avoid having to add new fields to the database we implement
a pseudo-challenge type "DNS-01-Wildcard" that can be used to hint
policy checks in subsequent layers (e.g. for CAA issuewild checking).
This challenge type is rendered as a normal DNS-01 challenge for API
consumers.