Remove background subprocess for DB migrations.
We started running our DB migrations in the background to speed up CI. However,
the semantics of subprocesses and `wait` mean that if a migration fails, the
overall `create_db.sh` doesn't fail. That means, for instance, tests continue to
run, and it's hard to find the resulting error.

This change runs the migrations in serial again so that we can catch such errors
more easily.