wildcard policy issuance allowing `*` in idents.
This commit reworks the initial wildcard policy based issuance support
to allow a "*." prefix for DNS type identifier values. We allow this
only **internal** to Boulder to maintain compliance with the ACME spec
that forbids wildcards in authorization identifiers. Before presenting
the authz/identifier to the user we strip the "*." prefix again.
Similarly before calling `va.PerformValidation` the RA strips the
wildcard prefix such that validation is performed for the base domain.

Allowing the wildcard character internal to boulder removes the
need to have a pseudo-challenge-type "DNS-01-Wildcard" to indicate that
an authz's identifier corresponds to the base domain of a wildcard
request. For CAA enforcement we have the information required in the
authz/identifier to make the right decisions.

Overall things are much cleaner. Thanks for the idea Roland!