Update Orders table migrations, add index for pendingOrder RL.
Since the `20170731115209_AddOrders.sql` migration hasn't been applied
yet there isn't any reason to modify the orders table structure in the
follow-up `20171006131843_FinalizeOrder.sql` migration. This commit
updates the `AddOrders` migration to not add a CSR field, that way the
`FinalizeOrder` migration doesn't need to drop/create it.

This commit also updates the Orders `KEY` that was previously on the
`registrationID` & `expires` fields to be on the `registrationID,
status, expires` fields so that the pendingOrdersByRegID rate limit
query will execute quickly.