Fix multiple issuewild record processing & add test.
Roland correctly spotted that the loop processing issuewild records was
returning false at the wrong level of indentation. This would prevent
issuance in a case where there were more than one issuewild records and
the VA CAA identity wasn't present in the first one we process. This
commit fixes that error and adds a test. The test failed before this
commit and passes afterwards.