Issue for wildcard name during v2 integration tests.
This commit updates `chisel2.py` to support DNS-01 challenges. The code
required was already there, but not "wired in" to the `auth_and_issue`
function. Along the way a bug with `get_chall` raising a string instead
of an exception was fixed.

This commit also updates `test/integration-test-v2.py` to issue
a certificate with a normal random domain name, and a random wildcard
domain name using DNS-01.