Issue for wildcard name during v2 integration tests. (#3301)
This commit updates `chisel2.py` to support DNS-01 challenges. The code
required was already there, but not "wired in" to the `auth_and_issue`
function. Along the way a bug with `get_chall` raising a string instead
of an exception was fixed.

This commit also updates `test/integration-test-v2.py` with two new tests
1) issuing a certificate with a normal random domain name, and a random wildcard
domain name using DNS-01.
2) issuing a certificate with a random domain name, and in the same certificate,
a wildcard for that domain name. This tests that two authorizations are created 
even though that base domain is the same.