Return empty Order from SA.GetOrderForNames, not nil.
gRPC cannot marshal a nil `*corepb.Order` in the case where
a `GetOrderForNames` RPC call to the SA finds no appropriate orders.
Instead this commit returns an empty order and the RA checks whether the
order has a non-nil ID to determine if there was an existing order to
reuse or not.