Handle Challenge "Errors" field for V2 API.
In ACME v1 challenges have an "error" field. In v2 challenges have an
"errors" field. This commit updates the `core/objects.go` `Challenge`
type to have both fields. We set the V2 "errors" field with the contents
of the V1 "error" field before clearing it in the V2
"prepChallengeForDisplay" function similar to how "uri" vs "url" is
handled.