Calculate order status by assoc. authz status.
This commit removes the stored Order status field and replaces it with
a value that is calculated on-the-fly by the SA when fetching an order,
based on the order's associated authorizations. Credit to Jacob for the
idea.

In summary:
* If any of the order's authorizations are invalid, the order is
  invalid.
* If any of the order's authorizations are deactivated, the order
  is deactivated.
* If any of the order's authorizations are pending, the order is
  pending.
* If all of the order's authorizations are valid, and there is
  a certificate serial, the order is valid.
* If all of the order's authorizations are valid, and we have
  began processing, but there is no certificate serial, the order is
  processing.
* If all of the order's authorizations are valid, and we haven't
  processing, then the order is pending waiting a finalization
  request.

This avoids having to explicitly update the order status when an
associated authorization changes status.