Optimize SA.CountPendingOrders implementation.
Since order status is now a calculated field populated based on the
individual statuses of the order's associated authorizations we can't
implement `SA.CountPendingAuthorizations`, the function used for the
pending order rate limit, as a simple "WHERE status = 'pending'" query.

Previous to this commit we worked around this by having
`SA.CountPendingAuthorizations` fetch all unexpired orders for
a registration, and then each of the individual authorizations for those
orders. This allowed calculating the status but was extremely costly!

This commit introduces an optimized version of
`SA.CountPendingAuthorizations` that works only on data from the
`orders` table. Instead of counting pending orders exactly, we count
orders that we know are not valid, processing or invalid. The leftover
unexpired orders must be deactivated or pending. This overcounts by
including deactivated orders in the pending orders count but is simple
to implement and will be performant.