ctpolicy: Add informational logs and don't cancel remaining submissions
Add a set of logs which will be submitted to but not relied on for theit SCTs,
this allows us to test submissions to a particular log or submit to a log which
is not yet approved by a browser/root program.

Also add a feature which stops cancellations of remaining submissions when racing
to get a SCT from a group of logs.