Add Retry-After header for authz/challenge poll
Some clients exhibit excessive retry patterns on authz and/or challenge polling,
but obey the Retry-After header. This change starts setting Retry-After to a
default value of 3 seconds, gated by a feature flag in case some clients break
badly in the presence of Retry-After (which we've never sent before).
If successful, this may reduce the amount of polling we receive from
misconfigured clients. We may then choose to refine the behavior further, for
instance by setting Retry-After to longer intervals based on the age or status
of the authorization. For instance, valid / invalid authorizations could have a
very large Retry-After, since they are expected to never change.