Run load-generator against V1 and V2 API in CI.
This commit adds short 15s runs of the load generator against the V1 and
V2 APIs during the three integration test runs (v1 config, v1
config-next, and v2).

The output latency from the load generator runs is validated to ensure
that all requests succeeded. We don't presently generate latency graphs
with the data.

A fix for a race condition in the load-generator code itself related to
HTTP status code tracking is included in this commit.